# backend/internal/api/gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// Defines values for Prefecture.
const (
	Empty Prefecture = "北海道"
	N1    Prefecture = "青森県"
	N10   Prefecture = "埼玉県"
	N11   Prefecture = "千葉県"
	N12   Prefecture = "東京都"
	N13   Prefecture = "神奈川県"
	N14   Prefecture = "新潟県"
	N15   Prefecture = "富山県"
	N16   Prefecture = "石川県"
	N17   Prefecture = "福井県"
	N18   Prefecture = "山梨県"
	N19   Prefecture = "長野県"
	N2    Prefecture = "岩手県"
	N20   Prefecture = "岐阜県"
	N21   Prefecture = "静岡県"
	N22   Prefecture = "愛知県"
	N23   Prefecture = "三重県"
	N24   Prefecture = "滋賀県"
	N25   Prefecture = "京都府"
	N26   Prefecture = "大阪府"
	N27   Prefecture = "兵庫県"
	N28   Prefecture = "奈良県"
	N29   Prefecture = "和歌山県"
	N3    Prefecture = "宮城県"
	N30   Prefecture = "鳥取県"
	N31   Prefecture = "島根県"
	N32   Prefecture = "岡山県"
	N33   Prefecture = "広島県"
	N34   Prefecture = "山口県"
	N35   Prefecture = "徳島県"
	N36   Prefecture = "香川県"
	N37   Prefecture = "愛媛県"
	N38   Prefecture = "高知県"
	N39   Prefecture = "福岡県"
	N4    Prefecture = "秋田県"
	N40   Prefecture = "佐賀県"
	N41   Prefecture = "長崎県"
	N42   Prefecture = "熊本県"
	N43   Prefecture = "大分県"
	N44   Prefecture = "宮崎県"
	N45   Prefecture = "鹿児島県"
	N46   Prefecture = "沖縄県"
	N5    Prefecture = "山形県"
	N6    Prefecture = "福島県"
	N7    Prefecture = "茨城県"
	N8    Prefecture = "栃木県"
	N9    Prefecture = "群馬県"
)

// Valid indicates whether the value is a known member of the Prefecture enum.
func (e Prefecture) Valid() bool {
	switch e {
	case Empty:
		return true
	case N1:
		return true
	case N10:
		return true
	case N11:
		return true
	case N12:
		return true
	case N13:
		return true
	case N14:
		return true
	case N15:
		return true
	case N16:
		return true
	case N17:
		return true
	case N18:
		return true
	case N19:
		return true
	case N2:
		return true
	case N20:
		return true
	case N21:
		return true
	case N22:
		return true
	case N23:
		return true
	case N24:
		return true
	case N25:
		return true
	case N26:
		return true
	case N27:
		return true
	case N28:
		return true
	case N29:
		return true
	case N3:
		return true
	case N30:
		return true
	case N31:
		return true
	case N32:
		return true
	case N33:
		return true
	case N34:
		return true
	case N35:
		return true
	case N36:
		return true
	case N37:
		return true
	case N38:
		return true
	case N39:
		return true
	case N4:
		return true
	case N40:
		return true
	case N41:
		return true
	case N42:
		return true
	case N43:
		return true
	case N44:
		return true
	case N45:
		return true
	case N46:
		return true
	case N5:
		return true
	case N6:
		return true
	case N7:
		return true
	case N8:
		return true
	case N9:
		return true
	default:
		return false
	}
}

// CreatePinRequest defines model for CreatePinRequest.
type CreatePinRequest struct {
	// AnonToken クライアントが生成・保持する匿名トークン（任意）。後からアカウントへ
	// claim するための識別子。サーバはそのまま保存し、地図には出さない。
	AnonToken *string `json:"anon_token,omitempty"`

	// City 市区町村（表示用テキスト）。サーバは municipality_code の正規名称で上書きする。
	City string `json:"city"`

	// Comment 一言コメント（任意）。
	Comment *string `json:"comment,omitempty"`

	// MunicipalityCode あいまい検索の候補から選択した全国地方公共団体コード（市区町村）。必須。
	// サーバはこのコードで実在・都道府県所属を検証し、座標をその境界内に生成する。
	// 未指定・実在しない・都道府県不一致のコードは 400 で拒否する。
	MunicipalityCode string `json:"municipality_code"`

	// Nickname 投稿者のニックネーム。
	Nickname string `json:"nickname"`

	// Prefecture 標準47都道府県
	Prefecture Prefecture `json:"prefecture"`

	// UtmCampaign キャンペーンの計測値（任意）。utm_campaign。
	UtmCampaign *string `json:"utm_campaign,omitempty"`

	// UtmMedium 流入媒体の計測値（任意）。utm_medium。
	UtmMedium *string `json:"utm_medium,omitempty"`

	// UtmSource 流入元の計測値（任意）。X 共有リンク等の utm_source。後から復元不可能なため投稿時に記録する。
	UtmSource *string `json:"utm_source,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Message ユーザー向けエラーメッセージ。
	Message string `json:"message"`
}

// Pin defines model for Pin.
type Pin struct {
	// City 市区町村（表示用）。seed 由来のピンでは空。
	City *string `json:"city,omitempty"`

	// Comment 一言コメント。seed 由来のピンでは空。
	Comment *string `json:"comment,omitempty"`

	// Lat Example: 33.56
	Lat float64 `json:"lat"`

	// Lng Example: 133.53
	Lng float64 `json:"lng"`

	// Nickname 投稿者のニックネーム。seed 由来のピンでは空。
	Nickname *string `json:"nickname,omitempty"`

	// Prefecture 標準47都道府県
	Prefecture Prefecture `json:"prefecture"`

	// Weight ヒートの重み。最小スライスでは件数で密度を表現するため通常 1。
	Weight *int `json:"weight,omitempty"`
}

// PinsResponse defines model for PinsResponse.
type PinsResponse struct {
	Pins []Pin `json:"pins"`

	// PrefectureCount ヒーロー指標。何都道府県に散らばっているか。
	PrefectureCount int `json:"prefecture_count"`

	// Total ピンの総数（投稿件数。同一ファンの連投も含む）。
	Total int `json:"total"`

	// UniqueFans ip_hash で重複排除した人数。画面の「人」表示にはこちらを使う
	// （total は投稿件数であり連投で水増しされるため）。
	UniqueFans int `json:"unique_fans"`
}

// Prefecture 標準47都道府県
type Prefecture string

// PrefectureStat defines model for PrefectureStat.
type PrefectureStat struct {
	// Count その都道府県のユニークファン数（ip_hashで重複排除）。
	Count int `json:"count"`

	// Prefecture 標準47都道府県
	Prefecture Prefecture `json:"prefecture"`
}

// GetPrefectureAtParams defines parameters for GetPrefectureAt.
type GetPrefectureAtParams struct {
	// Lat クリック地点の緯度。
	Lat float64 `form:"lat" json:"lat"`

	// Lng クリック地点の経度。
	Lng float64 `form:"lng" json:"lng"`
}

// PostApiPinsJSONRequestBody defines body for PostApiPins for application/json ContentType.
type PostApiPinsJSONRequestBody = CreatePinRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetApiPins ヒートマップ用のピン一覧を取得する
	// (GET /api/pins)
	GetApiPins(c *gin.Context)
	// PostApiPins ファンがピンを1件投稿する（認証なし）
	// (POST /api/pins)
	PostApiPins(c *gin.Context)
	// GetPrefectureAt クリック地点の都道府県とそのピン合計件数を返す
	// (GET /api/prefectures/at)
	GetPrefectureAt(c *gin.Context, params GetPrefectureAtParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetApiPins operation middleware
func (siw *ServerInterfaceWrapper) GetApiPins(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetApiPins(c)
}

// PostApiPins operation middleware
func (siw *ServerInterfaceWrapper) PostApiPins(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostApiPins(c)
}

// GetPrefectureAt operation middleware
func (siw *ServerInterfaceWrapper) GetPrefectureAt(c *gin.Context) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetPrefectureAtParams

	// ------------- Required query parameter "lat" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "lat", c.Request.URL.Query(), &params.Lat, runtime.BindQueryParameterOptions{Type: "number", Format: "double"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter lat: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "lng" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "lng", c.Request.URL.Query(), &params.Lng, runtime.BindQueryParameterOptions{Type: "number", Format: "double"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter lng: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetPrefectureAt(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/api/pins", wrapper.GetApiPins)
	router.POST(options.BaseURL+"/api/pins", wrapper.PostApiPins)
	router.GET(options.BaseURL+"/api/prefectures/at", wrapper.GetPrefectureAt)
}

type GetApiPinsRequestObject struct {
}

type GetApiPinsResponseObject interface {
	VisitGetApiPinsResponse(w http.ResponseWriter) error
}

type GetApiPins200JSONResponse PinsResponse

func (response GetApiPins200JSONResponse) VisitGetApiPinsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetApiPins500JSONResponse Error

func (response GetApiPins500JSONResponse) VisitGetApiPinsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type PostApiPinsRequestObject struct {
	Body *PostApiPinsJSONRequestBody
}

type PostApiPinsResponseObject interface {
	VisitPostApiPinsResponse(w http.ResponseWriter) error
}

type PostApiPins201JSONResponse Pin

func (response PostApiPins201JSONResponse) VisitPostApiPinsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type PostApiPins400JSONResponse Error

func (response PostApiPins400JSONResponse) VisitPostApiPinsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PostApiPins500JSONResponse Error

func (response PostApiPins500JSONResponse) VisitPostApiPinsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetPrefectureAtRequestObject struct {
	Params GetPrefectureAtParams
}

type GetPrefectureAtResponseObject interface {
	VisitGetPrefectureAtResponse(w http.ResponseWriter) error
}

type GetPrefectureAt200JSONResponse PrefectureStat

func (response GetPrefectureAt200JSONResponse) VisitGetPrefectureAtResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetPrefectureAt400JSONResponse Error

func (response GetPrefectureAt400JSONResponse) VisitGetPrefectureAtResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetPrefectureAt404JSONResponse Error

func (response GetPrefectureAt404JSONResponse) VisitGetPrefectureAtResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetPrefectureAt500JSONResponse Error

func (response GetPrefectureAt500JSONResponse) VisitGetPrefectureAtResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetApiPins ヒートマップ用のピン一覧を取得する
	// (GET /api/pins)
	GetApiPins(ctx context.Context, request GetApiPinsRequestObject) (GetApiPinsResponseObject, error)
	// PostApiPins ファンがピンを1件投稿する（認証なし）
	// (POST /api/pins)
	PostApiPins(ctx context.Context, request PostApiPinsRequestObject) (PostApiPinsResponseObject, error)
	// GetPrefectureAt クリック地点の都道府県とそのピン合計件数を返す
	// (GET /api/prefectures/at)
	GetPrefectureAt(ctx context.Context, request GetPrefectureAtRequestObject) (GetPrefectureAtResponseObject, error)
}

type StrictHandlerFunc func(ctx *gin.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictGinServerOptions struct {
	// RequestErrorHandlerFunc is called when a request cannot be parsed or
	// decoded. It is invoked for JSON bind failures, form parse/bind errors,
	// multipart reader errors, media type parse errors, missing multipart
	// boundaries, and request body read errors. The default returns 400.
	RequestErrorHandlerFunc func(ctx *gin.Context, err error)
	// HandlerErrorFunc is called when the application handler (or any
	// middleware wrapping it) returns a non-nil error. The default returns 500.
	HandlerErrorFunc func(ctx *gin.Context, err error)
	// ResponseErrorHandlerFunc is called when the response object fails to
	// serialize (Visit*Response returns an error) or when the handler returns
	// an unexpected response type. The default returns 500.
	ResponseErrorHandlerFunc func(ctx *gin.Context, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictGinServerOptions{
		RequestErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		},
		HandlerErrorFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
		ResponseErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictGinServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		}
	}
	if options.HandlerErrorFunc == nil {
		options.HandlerErrorFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictGinServerOptions
}

// GetApiPins operation middleware
func (sh *strictHandler) GetApiPins(ctx *gin.Context) {
	var request GetApiPinsRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetApiPins(ctx, request.(GetApiPinsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetApiPins")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetApiPinsResponseObject); ok {
		if err := validResponse.VisitGetApiPinsResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostApiPins operation middleware
func (sh *strictHandler) PostApiPins(ctx *gin.Context) {
	var request PostApiPinsRequestObject

	var body PostApiPinsJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(ctx, err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostApiPins(ctx, request.(PostApiPinsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostApiPins")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(PostApiPinsResponseObject); ok {
		if err := validResponse.VisitPostApiPinsResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetPrefectureAt operation middleware
func (sh *strictHandler) GetPrefectureAt(ctx *gin.Context, params GetPrefectureAtParams) {
	var request GetPrefectureAtRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetPrefectureAt(ctx, request.(GetPrefectureAtRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetPrefectureAt")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetPrefectureAtResponseObject); ok {
		if err := validResponse.VisitGetPrefectureAtResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}