# backend/openapi.yaml
openapi: 3.0.3
info:
  title: glutton_map API
  version: 0.1.0
  description: |
    最小ヒートマップ・スライスの契約。これが唯一の真実（spec-first）。
    Go サーバ型は oapi-codegen、フロント TS 型は openapi-typescript で本ファイルから生成する。

    注: 以下は JSON ではないため本契約（strict-server 生成）の対象外。素の Gin ルートとして
    cmd/server/main.go で登録する（internal/share, internal/outbound）:
      - GET /share          … X 共有用に OGP/Twitter Card 入り HTML を SSR で返す
      - GET /static/ogp.png … 静的 OGP 画像（image/png）
      - GET /out?to=<key>   … ホワイトリストされた公式 URL へ計測後 302 リダイレクト
servers:
  - url: http://localhost:8001
paths:
  /api/pins:
    get:
      operationId: getApiPins
      summary: ヒートマップ用のピン一覧を取得する
      responses:
        "200":
          description: ピン一覧と集計
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinsResponse"
        "500":
          description: サーバ内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postApiPins
      summary: ファンがピンを1件投稿する（認証なし）
      description: |
        座標はサーバが都道府県の重心+ゆらぎで生成する。クライアントから lat/lng は受け取らない。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePinRequest"
      responses:
        "201":
          description: 作成されたピン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "400":
          description: 入力が不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: サーバ内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/prefectures/at:
    get:
      operationId: getPrefectureAt
      summary: クリック地点の都道府県とそのピン合計件数を返す
      description: |
        地図クリックの座標(lat/lng)を受け取り、行政区域ポリゴンで都道府県を判定して
        その都道府県のピン合計件数を返す。地点がどの都道府県にも属さない(海上など)場合は 404。
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
          description: クリック地点の緯度。
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
          description: クリック地点の経度。
      responses:
        "200":
          description: 都道府県とピン合計件数
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrefectureStat"
        "400":
          description: パラメータが不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 地点に該当する都道府県がない(海上など)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: サーバ内部エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Pin:
      type: object
      required: [prefecture, lat, lng]
      properties:
        prefecture:
          $ref: "#/components/schemas/Prefecture"
        lat:
          type: number
          format: double
          example: 33.56
        lng:
          type: number
          format: double
          example: 133.53
        weight:
          type: integer
          default: 1
          description: ヒートの重み。最小スライスでは件数で密度を表現するため通常 1。
        nickname:
          type: string
          description: 投稿者のニックネーム。seed 由来のピンでは空。
        city:
          type: string
          description: 市区町村（表示用）。seed 由来のピンでは空。
        comment:
          type: string
          description: 一言コメント。seed 由来のピンでは空。
    CreatePinRequest:
      type: object
      required: [nickname, prefecture, city, municipality_code]
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 30
          description: 投稿者のニックネーム。
        prefecture:
          $ref: "#/components/schemas/Prefecture"
        city:
          type: string
          minLength: 1
          maxLength: 50
          description: 市区町村（表示用テキスト）。サーバは municipality_code の正規名称で上書きする。
        municipality_code:
          type: string
          minLength: 1
          description: |
            あいまい検索の候補から選択した全国地方公共団体コード（市区町村）。必須。
            サーバはこのコードで実在・都道府県所属を検証し、座標をその境界内に生成する。
            未指定・実在しない・都道府県不一致のコードは 400 で拒否する。
        comment:
          type: string
          maxLength: 200
          description: 一言コメント（任意）。
        anon_token:
          type: string
          maxLength: 64
          description: |
            クライアントが生成・保持する匿名トークン（任意）。後からアカウントへ
            claim するための識別子。サーバはそのまま保存し、地図には出さない。
        utm_source:
          type: string
          maxLength: 64
          description: 流入元の計測値（任意）。X 共有リンク等の utm_source。後から復元不可能なため投稿時に記録する。
        utm_medium:
          type: string
          maxLength: 64
          description: 流入媒体の計測値（任意）。utm_medium。
        utm_campaign:
          type: string
          maxLength: 64
          description: キャンペーンの計測値（任意）。utm_campaign。
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: ユーザー向けエラーメッセージ。
    PinsResponse:
      type: object
      required: [pins, prefecture_count, total, unique_fans]
      properties:
        pins:
          type: array
          items:
            $ref: "#/components/schemas/Pin"
        prefecture_count:
          type: integer
          description: ヒーロー指標。何都道府県に散らばっているか。
        total:
          type: integer
          description: ピンの総数（投稿件数。同一ファンの連投も含む）。
        unique_fans:
          type: integer
          description: |
            ip_hash で重複排除した人数。画面の「人」表示にはこちらを使う
            （total は投稿件数であり連投で水増しされるため）。
    PrefectureStat:
      type: object
      required: [prefecture, count]
      properties:
        prefecture:
          $ref: "#/components/schemas/Prefecture"
        count:
          type: integer
          description: その都道府県のユニークファン数（ip_hashで重複排除）。
    Prefecture:
      type: string
      description: 標準47都道府県
      enum:
        - 北海道
        - 青森県
        - 岩手県
        - 宮城県
        - 秋田県
        - 山形県
        - 福島県
        - 茨城県
        - 栃木県
        - 群馬県
        - 埼玉県
        - 千葉県
        - 東京都
        - 神奈川県
        - 新潟県
        - 富山県
        - 石川県
        - 福井県
        - 山梨県
        - 長野県
        - 岐阜県
        - 静岡県
        - 愛知県
        - 三重県
        - 滋賀県
        - 京都府
        - 大阪府
        - 兵庫県
        - 奈良県
        - 和歌山県
        - 鳥取県
        - 島根県
        - 岡山県
        - 広島県
        - 山口県
        - 徳島県
        - 香川県
        - 愛媛県
        - 高知県
        - 福岡県
        - 佐賀県
        - 長崎県
        - 熊本県
        - 大分県
        - 宮崎県
        - 鹿児島県
        - 沖縄県